backend/app/utils/image_processing.py
# backend/app/utils/image_processing.py
from PIL import Image
import numpy as np
import cv2

def process_image(image):
    img = Image.open(image)
    # Perform image processing if needed
    return img

def color_histogram(img):
    histogram = cv2.calcHist([np.array(img)], [0, 1, 2], None, [256, 256, 256], [0, 256, 0, 256, 0, 256])
    return histogram

def segmentation_mask(img):
    # Dummy segmentation for example
    mask = Image.new('L', img.size, 0)
    draw = ImageDraw.Draw(mask)
    draw.rectangle([50, 50, 200, 200], fill=255)
    return mask
