backend/requirements.txt
# backend/requirements.txt
flask
pandas
numpy
scikit-learn
opencv-python
Pillow
nltk
textblob
flask_pymongo
pydantic